+++ src/components/buttons/ButtonSgPref.tsx
@@ -38,7 +38,7 @@ const useStyles = createUseStyles<string>({
   sgPrefButton: {
     left: 'calc(60% + 6rem)',
     'html[data-noanime="false"] &': {
-      '&:hover, &.windowactive': {
+      '&:hover, &:focus, &.windowactive': {
         '& svg': {
           '& .el-1': {
             visibility: 'hidden',